requirements.txt
streamlit
pandas
numpy
PyPDF2
python-docx
nltk
scikit-learn
plotly
wordcloud
textstat
streamlit-option-menu
Pillow